+++ test/bargs.test.ts
@@ -8,92 +8,7 @@ import type { StringOption } from '../src/types.js';
 
 import { bargs, handle, map, merge } from '../src/bargs.js';
 import { opt, pos } from '../src/opt.js';
-
-/**
- * Custom error thrown when process.exit is mocked and called.
- */
-class MockExitError extends Error {
-  readonly exitCode: number;
-
-  constructor(exitCode: number) {
-    super(`process.exit(${exitCode}) was called`);
-    this.name = 'MockExitError';
-    this.exitCode = exitCode;
-  }
-}
-
-/**
- * Helper to capture stderr output and mock process.exit during tests. Returns
- * the captured output and exit code.
- *
- * @function
- */
-const withMockedExit = <T>(
-  fn: () => Promise<T> | T,
-): Promise<{
-  exitCode: number;
-  output: string;
-  result?: T;
-}> => {
-  const stderrWrites: string[] = [];
-  const originalStderrWrite = process.stderr.write.bind(process.stderr);
-  const originalExit = process.exit;
-
-  process.stderr.write = ((chunk: unknown) => {
-    stderrWrites.push(String(chunk));
-    return true;
-  }) as typeof process.stderr.write;
-
-  // Mock process.exit to throw instead of actually exiting
-  process.exit = ((code?: number) => {
-    throw new MockExitError(code ?? 0);
-  }) as typeof process.exit;
-
-  /**
-   * @function
-   */
-  const cleanup = () => {
-    process.stderr.write = originalStderrWrite;
-    process.exit = originalExit;
-  };
-
-  /**
-   * @function
-   */
-  const handleMockExit = (error: unknown) => {
-    if (error instanceof MockExitError) {
-      return {
-        exitCode: error.exitCode,
-        output: stderrWrites.join(''),
-        result: undefined,
-      };
-    }
-    throw error;
-  };
-
-  try {
-    const maybePromise = fn();
-    if (maybePromise instanceof Promise) {
-      return maybePromise
-        .then((result) => ({
-          exitCode: 0,
-          output: stderrWrites.join(''),
-          result,
-        }))
-        .catch(handleMockExit)
-        .finally(cleanup);
-    }
-    cleanup();
-    return Promise.resolve({
-      exitCode: 0,
-      output: stderrWrites.join(''),
-      result: maybePromise,
-    });
-  } catch (error) {
-    cleanup();
-    return Promise.resolve(handleMockExit(error));
-  }
-};
+import { withMockedExit } from './helpers/mock-exit.js';
 
 describe('bargs()', () => {
   it('creates a CLI builder', () => {
